gynecology_chatbot/chainlit_app/requirements.txt
# Chainlit
chainlit==1.0.101
aiohttp==3.9.1
python-dotenv==1.0.0

# Utilities
pydantic==2.4.2
